pyproject.toml
[tool.poetry]
name = "wishlist"
version = "0.1.0"
description = ""
authors = ["Mohsin-Ul-Islam <mohsinulislam180@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["all"], version = "^0.81.0"}
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^7"
isort = "^5.10.1"
mypy = "^0.971"
black = {version = "^22.8.0", allow-prereleases = true}
"testing.postgresql" = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
